LayoutTests/imported/w3c/web-platform-tests/webcodecs/reconfiguring-encoder.https.any_h264_avc-expected.txt

CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 800 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code
CONSOLE MESSAGE: Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code

Harness Error (FAIL), message = Error: assert_unreached: assert_equals: expected (number) 640 but got (undefined) undefined Reached unreachable code

PASS Reconfiguring encoder
